+++ AutoPatcher.Engine/ArtifactLocator/ArtifactLocatorException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace AutoPatcher.Engine.ArtifactLocator
+{
+    internal sealed class ArtifactLocatorException : Exception
+    {
+        public ArtifactLocatorException(string message) : base(message) { }
+    }
+}