+++ AutoPatcher.Engine/ArtifactLocator/IArtifactLocator.cs
@@ -1,0 +1,7 @@
+﻿namespace AutoPatcher.Engine.ArtifactLocator
+{
+    public interface IArtifactLocator
+    {
+        string GetPathFromAssemblyName(string assemblyName);
+    }
+}