+++ AutoPatcher.Engine/ArtifactLocator/IArtifactTypeMatcher.cs
@@ -1,0 +1,7 @@
+﻿namespace AutoPatcher.Engine.ArtifactLocator
+{
+    public interface IArtifactTypeMatcher
+    {
+        bool IsOfType(string filePath);
+    }
+}