+++ AutoPatcher.Engine/MSBuild/MSBuildImportException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace AutoPatcher.Engine.MSBuild
+{
+    internal sealed class MSBuildImportException : Exception
+    {
+        public MSBuildImportException(string message) : base(message) { }
+    }
+}